Requested forms are filled with the first and last name of the applicant
This adds a method, form_field_hash, to Resident that returns the names of
fields in a PDF mapped to what they should contain for this resident.

Our sample PDFs don't have a very consistent naming convention for their fields.
It'll take some thought to develop a naming convention for fields that's useful
and general.  For now, at least both of our sample forms contain "FirstName" and
"LastName".