Fill forms locally with pdf-forms
This gets rid of the dependencies on GSA-OCSIT/pdf-filler, saving us from
relying on web requests to, and the existence of, an outside-hosted web service
and PDF forms.  The app is now self-contained, and you can fill/download the
sample forms with no internet connection at all.

The two sample forms are in /public/forms.  The database seeds and test
fixtures are generated from those.