Create a global Dragoman object
This seems more efficient and cleaner than re-creating a Dragoman object every
time a resident wants to know something about a set of field names.

In the context of this app, Dragoman seems like a good candidate for a
singleton object, since all of the translation rules sent to Dragoman#learn
never change.

This also gives us a nicer place to store all of the translation rules, in
config/initializers/dragoman.rb.