Add Location metadata to PDFs
I'd like the PDFs to get the correct location attribute when seeded or
added, without having to code the addresses into seeds.rb.

PDFs support adding arbitrary metadata fields.  I added a field called
"Location" to both of the sample PDFs, with the value of the
corresponding building's address.  I should be able to extract that data
with pdftk during seeding.