Primitive Salesforce integration
Fixes #16, hooray!

Adds the ability to pull applicants from the Bread for the City Salesforce test
sandbox.  There's a "Pull from salesforce" button on the applicants
page.  If you press it, and the app is was provided with Salesforce
login information in config/databasedotcom.yml and in the environment, it will
sync applicants in the local database with applicants in salesforce.

Right now it only gets first and last names.  Adding more fields is
easy, but tedious, so I'm leaving that for another issue.

I got the login information for the BftC Salesforce test sandbox from Andrew
Lomax (alomax@breadforthecity.org).  If you want to help develop against
that test instance, ask Andrew for login credentials.

The code expects the salesforce password in the environment variable
SALESFORCE_PASSWORD.  That's not tracked in the repo.