Convert PeopleController to IdentityController
Editing an identity no longer depends on the session's
current_applicant.  This is a good thing.  Identities are now edited
based on the applicant in the URL, not the person number.

This is relevant to my sanity as I work on fixing the way we page through
the form.  Errors described in #172.