Re-initialize form fields on seeds
Fields should get initialized when uploaded or added to the system.
However, with all the updates we've been making to the seeds, some forms
on on my workstation have a PDF, but no fields are scanned.  It should
be kind of a one-time thing, but for convenience, the seeds file will
re-normalize this.