Replace tabs with accordion menu
Fixes codefordc/districthousing#258.

Updates the edit views to replace the section tabs. Progress
bar moved into the breadcrumb trail. New CSS is added
for the menu items. Save/Delete buttons moved because
with many items (such as 7 sources of income), the buttons
were all the way at the bottom of the page.