Fixes a problem if the container is stopped and restarted
If a container is stopped or killed, it is equivalent to sending
SIGTERM or SIGKILL to the rails process.  However, the container
filesystem will still have all the server pids.

This adds an init script which runs and cleans up the tmp
directory so that a container restart acts like a fresh install.